Use only on instance of MongoClient
Implement DB as a singleton and use only one instance of MongoClient. MongoClient as to be initialize during
the startup and every time the configuration is submitter. It cannot be initialize via start() method, as BuildAnalysis descriptor is not
fully loaded yet, so it has to be done in postInitialize() method.